agentkit/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
web3>=6.15.1
langchain>=0.1.5
langchain-community>=0.0.13
langchain-core>=0.1.18
langchain-google-genai>=0.0.5
pydantic>=2.5.3
google-generativeai>=0.3.2
